simulate_colab_results.py
#!/usr/bin/env python3
"""
Script para simular resultados de Google Colab
(Para probar el flujo completo antes de usar Colab real)
"""

import json
import os
import time
import random
import numpy as np
from datetime import datetime
from typing import Dict, List

def generate_simulated_results() -> Dict:
    """Genera resultados simulados que imitan los de Colab"""
    
    # Configuración simulada
    models = ['multi-qa-mpnet-base-dot-v1', 'all-MiniLM-L6-v2', 'ada', 'e5-large-v2']
    num_questions = 500
    
    results = {}
    
    # Generar resultados para cada modelo
    for model_name in models:
        print(f"Generando resultados simulados para {model_name}...")
        
        # Generar métricas individuales (simulando evaluación real)
        individual_metrics = []
        for _ in range(num_questions):
            metrics = {
                'precision': random.uniform(0.6, 0.95),
                'recall': random.uniform(0.55, 0.90),
                'f1': random.uniform(0.58, 0.92),
                'map': random.uniform(0.50, 0.85),
                'mrr': random.uniform(0.55, 0.90),
                'ndcg': random.uniform(0.60, 0.95)
            }
            individual_metrics.append(metrics)
        
        # Calcular métricas promedio
        avg_metrics = {}
        for metric in ['precision', 'recall', 'f1', 'map', 'mrr', 'ndcg']:
            values = [m[metric] for m in individual_metrics]
            avg_metrics[f'avg_{metric}'] = np.mean(values)
            avg_metrics[f'std_{metric}'] = np.std(values)
        
        # Simular tiempos de procesamiento
        processing_time = random.uniform(30, 120)  # 30s a 2min por modelo
        
        results[model_name] = {
            'model_name': model_name,
            'avg_metrics': avg_metrics,
            'individual_metrics': individual_metrics,
            'total_questions': num_questions,
            'processing_time_seconds': processing_time,
            'gpu_used': True,
            'evaluation_time': datetime.now().isoformat(),
            'colab_session': True
        }
    
    # Compilar resultados finales
    total_time = sum(r['processing_time_seconds'] for r in results.values())
    
    final_results = {
        'config': {
            'num_questions': num_questions,
            'selected_models': models,
            'generative_model_name': 'llama-3.3-70b',
            'top_k': 10,
            'use_llm_reranker': True,
            'batch_size': 50,
            'evaluate_all_models': True,
            'evaluation_type': 'cumulative_metrics',
            'timestamp': datetime.now().strftime("%Y-%m-%d %H:%M:%S")
        },
        'results': results,
        'execution_summary': {
            'total_time_seconds': total_time,
            'questions_processed': num_questions,
            'models_evaluated': len(results),
            'gpu_used': True,
            'timestamp': datetime.now().isoformat(),
            'colab_session': True,
            'success_rate': 1.0
        }
    }
    
    return final_results

def save_simulated_results():
    """Guarda resultados simulados en la estructura esperada"""
    
    # Directorios
    base_dir = "/Users/haroldgomez/Documents/ProyectoTituloMAgister/SupportModel/simulated_drive"
    results_dir = f"{base_dir}/results"
    
    # Crear directorios
    os.makedirs(results_dir, exist_ok=True)
    
    # Generar resultados
    print("🎲 Generando resultados simulados de Google Colab...")
    evaluation_results = generate_simulated_results()
    
    # Nombres de archivos con timestamp
    timestamp = int(time.time())
    json_filename = f"cumulative_results_{timestamp}.json"
    csv_filename = f"results_summary_{timestamp}.csv"
    
    # Guardar JSON completo
    json_path = f"{results_dir}/{json_filename}"
    with open(json_path, 'w', encoding='utf-8') as f:
        json.dump(evaluation_results, f, indent=2, ensure_ascii=False)
    
    print(f"✅ Resultados JSON guardados: {json_path}")
    
    # Crear CSV resumen
    import pandas as pd
    
    csv_data = []
    for model_name, model_results in evaluation_results['results'].items():
        metrics = model_results['avg_metrics']
        csv_data.append({
            'Model': model_name,
            'Precision': f"{metrics['avg_precision']:.4f}",
            'Recall': f"{metrics['avg_recall']:.4f}",
            'F1_Score': f"{metrics['avg_f1']:.4f}",
            'MAP': f"{metrics['avg_map']:.4f}",
            'MRR': f"{metrics['avg_mrr']:.4f}",
            'NDCG': f"{metrics['avg_ndcg']:.4f}",
            'Time_s': f"{model_results['processing_time_seconds']:.2f}"
        })
    
    df = pd.DataFrame(csv_data)
    csv_path = f"{results_dir}/{csv_filename}"
    df.to_csv(csv_path, index=False)
    
    print(f"✅ Resumen CSV guardado: {csv_path}")
    
    # Crear archivo de status
    status_data = {
        'status': 'completed',
        'timestamp': datetime.now().isoformat(),
        'results_file': json_filename,
        'summary_file': csv_filename,
        'models_evaluated': len(evaluation_results['results']),
        'questions_processed': evaluation_results['execution_summary']['questions_processed'],
        'total_time_seconds': evaluation_results['execution_summary']['total_time_seconds'],
        'gpu_used': evaluation_results['execution_summary']['gpu_used']
    }
    
    status_path = f"{base_dir}/evaluation_status.json"
    with open(status_path, 'w', encoding='utf-8') as f:
        json.dump(status_data, f, indent=2)
    
    print(f"✅ Status guardado: {status_path}")
    
    # Mostrar resumen
    print("\n📊 RESUMEN DE RESULTADOS SIMULADOS:")
    print("=" * 50)
    
    # Ordenar por F1-Score
    ranking = sorted(evaluation_results['results'].items(), 
                    key=lambda x: x[1]['avg_metrics']['avg_f1'], 
                    reverse=True)
    
    for i, (model_name, model_results) in enumerate(ranking, 1):
        metrics = model_results['avg_metrics']
        print(f"{i}. {model_name}")
        print(f"   F1-Score: {metrics['avg_f1']:.4f}")
        print(f"   Precision: {metrics['avg_precision']:.4f}")
        print(f"   Recall: {metrics['avg_recall']:.4f}")
        print(f"   Tiempo: {model_results['processing_time_seconds']:.2f}s")
    
    print(f"\n🎉 Simulación completada!")
    print(f"📁 Archivos disponibles en: {results_dir}")
    print(f"🔄 Ve a Streamlit y presiona 'Verificar Estado' para ver los resultados")

def main():
    """Función principal"""
    print("🚀 Simulador de Resultados de Google Colab")
    print("=" * 50)
    print("Este script simula los resultados que generaría Google Colab")
    print("para probar el flujo completo del sistema.\n")
    
    save_simulated_results()

if __name__ == "__main__":
    main()