requirements.txt
openai==1.93.0
chromadb==0.5.23
python-dotenv==1.1.1
streamlit==1.46.1
scikit-learn==1.7.0
torch==2.2.2
numpy==1.26.4
sentence-transformers==5.0.0
plotly==6.2.0
weasyprint==63.1
kaleido==0.2.1
bert-score==0.3.13
rouge-score==0.1.2
google-generativeai==0.8.5
transformers==4.44.0
accelerate==0.32.1
bitsandbytes==0.43.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-api-python-client==2.175.0